+++ packages/packages/apps/onboarding/src/components/ui/split-layout.tsx
@@ -22,7 +22,7 @@ const LAYOUT_TRANSITION_MS = 300;
 const LEFT_PANEL_CONTENT_WIDTH = 386;
 const LEFT_PANEL_PADDING_X = 80;
 const LEFT_PANEL_PADDING_TOP = 40;
-const LEFT_PANEL_GAP = 60;
+const LEFT_PANEL_GAP = 32;
 const IMAGE_MIN_WIDTH = 464;
 const CONTENT_IMAGE_MIN_GAP = 80;
